+++ src/gufo/thor/labs/base.py
@@ -100,6 +100,7 @@ class BaseLab(object):
 
     name: str
     image: str
+    default_version: Optional[str] = None
     docker_console_args: Optional[DockerConsoleArgs] = None
 
     def get_compose_config(
@@ -113,6 +114,7 @@ class BaseLab(object):
             "tty": True,
             "stdin_open": True,
             "hostname": node_config.name,
+            "cap_add": ["NET_ADMIN", "SYS_ADMIN"],
         }
         volumes = self.get_compose_volumes(config, lab_config, node_config)
         if volumes:
@@ -128,6 +130,8 @@ class BaseLab(object):
         """Generate compose image name."""
         if node_config.version:
             return f"{self.image}:{node_config.version}"
+        if self.default_version:
+            return f"{self.image}:{self.default_version}"
         return self.image
 
     def get_compose_networks(